D-12784 This uses the PR that originally solved encoding issue and fixes tests
However, the Meta queries no longer work, so need to either decide to NOT encode
anything and change a bunch of tests to not expect encoding by the time the
BuildResult is built OR to make the Meta queries get directly encoded, since they don't
utilize BuildResult.

Also, not sure how to get log4j working at runtime. Argh.